refactor(codex): split config TOML facade
