+++ src/cli/install-codex/codex-config-features.ts
@@ -1,0 +1,7 @@
+import { appendBlock, findTomlSection, replaceOrInsertSetting } from "./toml-section-editor"
+
+export function ensureFeatureEnabled(config: string, featureName: string): string {
+  const section = findTomlSection(config, "features")
+  if (!section) return appendBlock(config, `[features]\n${featureName} = true\n`)
+  return replaceOrInsertSetting(config, section, featureName, "true")
+}
